+++ src/api/v1/utils/HttpError.ts
@@ -1,5 +1,9 @@
 export default class HttpError extends Error {
-  constructor(private _message = 'Internal Server Error', private _code = 500) {
+  constructor(
+    private _message = 'Something went wrong.',
+    private _code = 500,
+    private _status = 'INTERNAL SERVER ERROR'
+  ) {
     super(_message);
   }
 
@@ -9,5 +13,9 @@ export default class HttpError extends Error {
 
   get code(): number {
     return this._code;
+  }
+
+  get status(): string {
+    return this._status;
   }
 }