+++ src/config/docs/basicInfo.ts
@@ -3,7 +3,7 @@ export default {
   info: {
     title: 'FoodForYou',
     description: 'Food Delivery App',
-    version: '1.0.0-alpha.2',
+    version: '1.0.0-rc',
     contact: {
       name: 'Florian Hundegger',
       email: 'flo.hundegger@gmail.com',