+++ src/config/docs/endpoints/auth.ts
@@ -57,14 +57,7 @@ export default {
           content: {
             'application/json': {
               schema: {
-                type: 'object',
-                properties: {
-                  message: {
-                    type: 'string',
-                    example: 'Internal Server Error',
-                    description: 'error message'
-                  }
-                }
+                $ref: '#/components/schemas/HttpError'
               }
             }
           }