+++ CHANGELOG.md
@@ -1,20 +1,44 @@
 # CHANGELOG
+
 All notable changes to this project are documented in this file following the [Keep a CHANGELOG](https://keepachangelog.com/en/1.0.0/) conventions.
 
 ## [Unreleased]
+
 ### Added
+
 - Implement Payments
+
+## [1.0.0-rc] - 2022-07-06
+
+### Added
+
+- Implement Payments
+- Added status propery on HttpError
+
 ### Changed
-- refactor code
+
+- Changed fail responses
+- Implement DIJ 
+
+### Security
+
+- Updated nodemon v2.0.18 -> 2.0.19
 
 ## [1.0.0-alpha.2] - 2022-06-26
+
 ### Added
+
 - Automatically change total price / status of order when updating.
+
 ### Changed
+
 - Changed docs structure
 - Refactored tests
+
 ### Fixed
+
 - Fixed validation bugs
 
 ## [1.0.0-alpha.1] - 2022-06-23
+
 Initial Version.